sumup/pre-process-data/requirements.txt
google-cloud-bigquery
pandas
numpy
google-auth
google-auth-oauthlib

